Simulate SNP for all subjects

The function creates a bigsnpr object, saves it in a rds file. and fills the file backed matrix generated with geneotypes for each subject.

Usage

G_func_simple(filename, MAF, N = 1e+05, M = 1e+05, block_size = 1000)

Arguments

filename

Filename for file backed matrix(FBM) and rds file.

MAF

A vector containing minor allele frequencies.

N

Amount of subjects.

block_size

Size of FBM to be processed in each iteration.

n_sib

Amount of siblings.

Value

FBM containing simulated SNPs for all subjects.